refactor: rename trial_utils.py → link_utils.py
The module now serves both Trials and Articles, so the trial-specific
filename was misleading. Renamed the file, updated all 11 import
statements, renamed the companion test file (test_trial_utils →
test_link_utils), and updated the one inline comment in models.py.
No behaviour change.
