+++ django/gregory/management/commands/feedreader_trials.py
@@ -7,7 +7,7 @@ from django.utils import timezone
 from gregory.classes import ClinicalTrial, EUTrialParser
 from gregory.functions import remove_utm
 from gregory.models import Trials, Sources
-from gregory.utils.trial_utils import identifiers_conflict, merge_links, canonical_link
+from gregory.utils.link_utils import identifiers_conflict, merge_links, canonical_link
 import feedparser
 import pytz
 import requests
