+++ django/gregory/management/commands/feedreader_trials_ctgov.py
@@ -24,7 +24,7 @@ from django.db.models import Q
 from django.utils import timezone
 from gregory.classes import ClinicalTrialsGovAPI, ClinicalTrial
 from gregory.models import Trials, Sources
-from gregory.utils.trial_utils import identifiers_conflict, merge_links, canonical_link
+from gregory.utils.link_utils import identifiers_conflict, merge_links, canonical_link
 import pytz
