+++ django/gregory/utils/link_utils.py
